Fix slice logic and move dispatch into component
